Rename components and references to have the same for each component and improve code
